requirements.txt
django
django-cleanup
django-environ
djangorestframework
django-filter
mysqlclient~=1.4.6
pillow
